Module tensorflow_proxy::grpc

Structs

BackendConnection

Connection to the prediction service that can be used for creating new clients.

BytesCodec

A Codec for encoding and decoding byte vectors to and from Tonic buffers.

BytesDecoder

Passthrough decoder for decoding a tonic::codec::DecodeBuf to a byte vector.

BytesEncoder

Passthrough encoder for encoding a byte vector to an tonic::codec::EncodeBuf.

FutureSyncWrapper

Utility wrapper to implement Sync for Futures by wrapping the inner Future in a SyncWrapper.

Functions

handle_request