Module tensorflow_proxy::grpc::proto

Modules

attr_value

Nested message and enum types in AttrValue.

collection_def

Nested message and enum types in CollectionDef.

feature

Nested message and enum types in Feature.

full_type_def

Nested message and enum types in FullTypeDef.

function_def

Nested message and enum types in FunctionDef.

function_spec

Nested message and enum types in FunctionSpec.

meta_graph_def

Nested message and enum types in MetaGraphDef.

node_def

Nested message and enum types in NodeDef.

op_def

Nested message and enum types in OpDef.

resource_handle_proto

Nested message and enum types in ResourceHandleProto.

saved_object

Nested message and enum types in SavedObject.

saver_def

Nested message and enum types in SaverDef.

serving

structured_value

Nested message and enum types in StructuredValue.

tensor_info

Nested message and enum types in TensorInfo.

tensor_shape_proto

Nested message and enum types in TensorShapeProto.

trackable_object_graph

Nested message and enum types in TrackableObjectGraph.

type_spec_proto

Nested message and enum types in TypeSpecProto.

Structs

AssetFileDef

An asset file def for a single file or a set of sharded files with the same name.

AttrValue

Protocol buffer representing the value for an attr used to configure an Op. Comment indicates the corresponding attr type. Only the field matching the attr type may be filled.

BoundedTensorSpecProto

A protobuf to represent tf.BoundedTensorSpec.

BytesList

LINT.IfChange Containers to hold repeated fundamental values.

CapturedTensor

CollectionDef

CollectionDef should cover most collections. To add a user-defined collection, do one of the following:

DictValue

Represents a Python dict keyed by str. The comment on Unicode from Value.string_value applies analogously.

Example

Feature

Containers for non-sequential data.

FeatureList

Containers for sequential data.

FeatureLists

Features

FloatList

FullTypeDef

Highly experimental and very likely to change. This encoding uses tags instead of dedicated messages for regularity. In particular the encoding imposes no restrictions on what the parameters of any type should be, which in particular needs to be true for type symbols.

FunctionDef

A function can be instantiated when the runtime can bind every attr with a value. When a GraphDef has a call to a function, it must have binding for every attr defined in the signature.

FunctionDefLibrary

A library is a set of named functions.

FunctionSpec

Represents FunctionSpec used in Function. This represents a function that has been wrapped as a TensorFlow Function.

GradientDef

GradientDef defines the gradient function of a function defined in a function library.

GraphDef

Represents the graph of operations

Int64List

ListValue

Represents a Python list.

MetaGraphDef

NOTE: This protocol buffer is evolving, and will go through revisions in the coming months.

NameAttrList

A list of attr names and their values. The whole list is attached with a string name. E.g., MatMul[T=float].

NamedTupleValue

Represents Python’s namedtuple.

NodeDef

NoneValue

Represents None.

OpDef

Defines an operation. A NodeDef in a GraphDef specifies an Op by using the “op” field which should match the name of a OpDef. LINT.IfChange

OpDeprecation

Information about version-dependent deprecation of an op

OpList

A collection of OpDefs

PairValue

Represents a (key, value) pair.

RegisteredGradient

RegisteredGradient stores a gradient function that is registered in the gradients library and used in the ops of a function in the function library. Unlike GradientDef, these gradients are identified by op type, and not directly linked to any function.

RegisteredSaver

ResourceHandleProto

Protocol buffer representing a handle to a tensorflow resource. Handles are not valid across executions, but can be serialized back and forth from within a single run.

SaveSliceInfoDef

SaveableObject

SavedAsset

A SavedAsset points to an asset in the MetaGraph.

SavedBareConcreteFunction

SavedConcreteFunction

Stores low-level information about a concrete function. Referenced in either a SavedFunction or a SavedBareConcreteFunction.

SavedConstant

SavedFunction

A function with multiple signatures, possibly with non-Tensor arguments.

SavedObject

SavedObjectGraph

SavedResource

A SavedResource represents a TF object that holds state during its lifetime. An object of this type can have a reference to a: create_resource() and an initialize() function.

SavedUserObject

A SavedUserObject is an object (in the object-oriented language of the TensorFlow program) of some user- or framework-defined class other than those handled specifically by the other kinds of SavedObjects.

SavedVariable

Represents a Variable that is initialized by loading the contents from the checkpoint.

SaverDef

Protocol buffer representing the configuration of a Saver.

SequenceExample

SignatureDef

SignatureDef defines the signature of a computation supported by a TensorFlow graph.

StructuredValue

StructuredValue represents a dynamically typed value representing various data structures that are inspired by Python data structures typically used in TensorFlow functions as inputs and outputs.

TensorInfo

Information about a Tensor necessary for feeding or retrieval.

TensorProto

Protocol buffer representing a tensor.

TensorShapeProto

Dimensions of a tensor.

TensorSpecProto

A protobuf to represent tf.TensorSpec.

TrackableObjectGraph

TupleValue

Represents a Python tuple.

TypeSpecProto

Represents a tf.TypeSpec

VariableDef

Protocol buffer representing a Variable.

VariantTensorDataProto

Protocol buffer representing the serialization format of DT_VARIANT tensors.

VersionDef

Version information for a piece of serialized data

Enums

DataType

(== suppress_warning documentation-presence ==) LINT.IfChange

FullTypeId

Experimental. Represents the complete type information of a TensorFlow value.

VariableAggregation

Indicates how a distributed variable will be aggregated.

VariableSynchronization

Indicates when a distributed variable will be synced.