Module tensorflow_proxy::grpc::proto::serving

Modules

inference_result

Nested message and enum types in InferenceResult.

input

Nested message and enum types in Input.

model_spec

Nested message and enum types in ModelSpec.

prediction_service_client

Generated client implementations.

Structs

Class

A single class.

ClassificationRequest

ClassificationResponse

ClassificationResult

Contains one result per input example, in the same order as the input in ClassificationRequest.

Classifications

List of classes for a single item (tensorflow.Example).

ExampleList

Specifies one or more fully independent input Examples. See examples at: https://github.com/tensorflow/tensorflow/blob/master/tensorflow/core/example/example.proto

ExampleListWithContext

Specifies one or more independent input Examples, with a common context Example.

GetModelMetadataRequest

GetModelMetadataResponse

InferenceResult

Inference result, matches the type of request or is an error.

InferenceTask

Inference request such as classification, regression, etc…

Input

ModelSpec

Metadata for an inference request such as the model name and version.

MultiInferenceRequest

Inference request containing one or more requests.

MultiInferenceResponse

Inference request containing one or more responses.

PredictRequest

PredictRequest specifies which TensorFlow model to run, as well as how inputs are mapped to tensors and how outputs are filtered before returning to user.

PredictResponse

Response for PredictRequest on successful run.

Regression

Regression result for a single item (tensorflow.Example).

RegressionRequest

RegressionResponse

RegressionResult

Contains one result per input example, in the same order as the input in RegressionRequest.

SignatureDefMap

Message returned for “signature_def” field.